fix: don't check pull_request_target workflows
This trigger is designed to NOT pull workflow files from the posted PR
(in contrast to pull_request), so it's safe to skip it.

https://docs.github.com/en/actions/writing-workflows/choosing-when-your-workflow-runs/events-that-trigger-workflows#pull_request_target

To quote,

```
This event runs in the context of the base of the pull request, rather
than in the context of the merge commit, as the pull_request event does.
This prevents execution of unsafe code from the head of the pull request
that could alter your repository or steal any secrets you use in your
workflow. This event allows your workflow to do things like label or
comment on pull requests from forks. Avoid using this event if you need
to build or run code from the pull request.
```

Note: For these targets, you pull PR via a separate checkout step that
targets the PR head; so the PR code is pulled by the (main head)
workflow itself.

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>