feat: use `uv pip compile` for update-constraints
pip-tools are very slow. What takes it 10+ minutes to resolve all
dependencies can be done in seconds with `uv`. This is especially
problematic when we attempt to run `update-constraints` against a repo
with a huge dependency tree (like instructlab/instructlab) which may
even time out in gate because of how slow pip-compile is.

`uv pip compile` is, largely, a drop-in replacement for `pip-compile`,
so the transition is very simple.

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>